Move CompartmentGroup to compartments.py
